+++ activerecord/lib/active_record/connection_adapters/postgresql/column.rb
@@ -20,6 +20,10 @@ module ActiveRecord
         end
         alias :array? :array
 
+        def enum?
+          type == :enum
+        end
+
         def sql_type
           super.delete_suffix("[]")
         end